finished ShipMoveCmdEventHandler function and modified unsigned int in TransactionEvents to doubles to match Ship and Port classes

+++ code/include/TransactionCheckEvent.hpp
@@ -13,7 +13,7 @@ public:
 
   TransactionCheckEvent();
 
-  TransactionCheckEvent(ActorId ship_id, ActorId port_id, unsigned int ship_gold, unsigned int ship_rum, unsigned int port_rum);
+  TransactionCheckEvent(ActorId ship_id, ActorId port_id, double ship_gold, double ship_rum, double port_rum);
 
   ~TransactionCheckEvent();
 
@@ -32,22 +32,22 @@ public:
   ActorId getPortId() const;
 
   /* Ship gold setter */
-  void setShipGold(unsigned int x);
+  void setShipGold(double x);
 
   /* Ship gold getter */
-  unsigned int getShipGold() const;
+  double getShipGold() const;
 
   /* Ship rum setter */
-  void setShipRum(unsigned int y);
+  void setShipRum(double y);
 
   /* Ship rum getter */
-  unsigned int getShipRum() const;
+  double getShipRum() const;
   
   /* Port rum setter */
-  void setPortRum(unsigned int y);
+  void setPortRum(double y);
 
   /* Port rum getter */
-  unsigned int getPortRum() const;
+  double getPortRum() const;
 
   /* TransactionCheckEvent's event type */
   static const EventType event_type;
@@ -58,9 +58,9 @@ private:
   ActorId ship_id, port_id;
 
   /* Rum and gold of actors associated with this event */
-  unsigned int ship_gold;
-  unsigned int ship_rum;
-  unsigned int port_rum;
+  double ship_gold;
+  double ship_rum;
+  double port_rum;
 };
 
 inline EventType TransactionCheckEvent::getEventType() const
@@ -88,32 +88,32 @@ inline ActorId TransactionCheckEvent::getPortId() const
   return port_id;
 }
 
-inline void TransactionCheckEvent::setShipGold(unsigned int x)
+inline void TransactionCheckEvent::setShipGold(double x)
 {
   this->ship_gold = x;
 }
 
-inline unsigned int TransactionCheckEvent::getShipGold() const
+inline double TransactionCheckEvent::getShipGold() const
 {
   return ship_gold;
 }
 
-inline void TransactionCheckEvent::setShipRum(unsigned int y)
+inline void TransactionCheckEvent::setShipRum(double y)
 {
   this->ship_rum = y;
 }
 
-inline unsigned int TransactionCheckEvent::getShipRum() const
+inline double TransactionCheckEvent::getShipRum() const
 {
   return ship_rum;
 }
 
-inline void TransactionCheckEvent::setPortRum(unsigned int y)
+inline void TransactionCheckEvent::setPortRum(double y)
 {
   this->port_rum= y;
 }
 
-inline unsigned int TransactionCheckEvent::getPortRum() const
+inline double TransactionCheckEvent::getPortRum() const
 {
   return port_rum;
 }
